Complete basic web frontend
Currently very crude and experimental.  This is one giant mega-commit
(shame on me).

The code largely still needs to be commented (bigger shame on me).

The biggest design oddity to note is that data persistence is currently
achieved with `pickle` (dumped to the server's local filesystem), rather
than a database.  This choice was made mainly out of curiosity, to see
how it would work out, having never worked with pickle.  At this point,
I am definitely leaning towards converting the objects to more typical
database models, if the project goes much further.

Also due to this choice, there is no multi-user capability right now.
If multiple users try to access the same instance of this app, without
knowing about each other, confusion will certainly ensue!

Game save-and-load "exists", in that your game is automatically saved
and loaded on each page reload.  But if you lose/exit your browser window
and revisit the home page, your only real option at the moment (without
magic special knowledge) is to delete your game and start over.

Error handling is near-nonexistent.  The app crashes easily because the
frontend neither prevents invalid data entry by the user very well, nor
handles exceptions raised by the backend.

The UI is exceedingly clunky, and entirely un-styled.  The worst
offender currently, the Game Status display, is nearly unreadable; it's
little more than a debugging info dump of the game state.  There are
many major UX improvements that could or should be made to the input
forms, as well.

BUT!  If you are a highly savvy and well-behaved user, all the
functionality needed to track a game of Clue and let the computer solve
it for you automatically, is now present.