Improve docs and add general querying mechanism
Continue adding and improving docstrings for backend methods.

Add a querying mechanism, ClueRelationFilter, to hopefully help simplify
the code of the Game class, and allow external querying, e.g. by the
frontend.  (This is really stuff that would come for "free" if we
converted to SQL-based models, but that's not my goal right now.)