Reorganize backend Game deduction logic
Rewrite querying in the Game class to use new ClueRelationFilter
querying mechanism.  The goal is to simplify the querying code and
reduce the need for custom assemblages of for loops and if statements.

Making use of the new querying power, move all of the querying-related
logic *out* of the Have and Show classes themselves (aside from a custom
__contains__ method on each).  Both those classes are thus greatly
simplified.

Also create private helper methods to encapsulate most of the deductions
that may need to be made when recording Haves or Shows.  I couldn't find
a way to DRY the code out much by doing this, but at least it breaks
methods into bite-sized chunks and hopefully improves readability.