Refactor backend classes and slightly change API
Code refactored around ClueRelationType enumeration, eliminating need
for custom classes per relation type.  The main reason for the refactor
is to make the code clearer and more explicit, and to work towards maybe
slowly DRYing it out a bit, and finally, to enable us to perhaps
eventually clarify and unify the API a bit better.  We do sacrifice the
natural complementarity of the Have and Pass relations, but the extra
code resulting from this is pretty negligible (and maybe we'll discover
a way to effectively add it back in within the new structure anyway).

The refactor results in an API change: record_have becomes
record_have_pass, and changes its parameters (without changing their
meaning).

Due to this API change, the frontend is updated accordingly.  Similarly,
the tests are updated (and obsolete ones discarded).

Besides the API change, querying syntax is significantly improved by
adding custom operators to ClueRelationFilter class.  This new
operator-based syntax is now used throughout the Game class.  Tests
again are updated accordingly.

Finally, the documentation is also updated to reflect recent changes.

This should provide a more solid base for future work.