Refactor Game class and update API
The main goal of this refactoring is to improve the API usability,
consistency, and clarity, while keeping the internal code DRY and
somewhat modular.

The Game implementation is also simplified by removing the notion of a
"myself" versus other players.  Tracking which player represents the
user, where it may be desired, is now delegated to clients of the API.
This does lead to a minor regression currently in the web frontend,
where drop-downs for players no longer omit the player's own self, even
when it would be appropriate to do so.  I consider this an insignificant
issue, and if needed it should be fixable in various other ways without
needing to re-complicate the Game class.

The frontend, tests, and docs are updated accordingly.